Removed experimental "master merging" from CombineVariants
git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@5138 348d0f76-0448-11de-a6fe-93d51630548a

+++ java/src/org/broadinstitute/sting/oneoffprojects/walkers/phasing/ComparePhasingToTrioPhasingNoRecombinationWalker.java
@@ -497,7 +497,7 @@ class CompareTrioAndPhasingTracks {
         out.println("");
     }
 
-    private static String percentString(double numerator, double denominator) {        
+    private static String percentString(int numerator, int denominator) {
         int NUM_DECIMAL_PLACES = 1;
         String percent = new Formatter().format("%." + NUM_DECIMAL_PLACES + "f", MathUtils.percentage(numerator, denominator)).toString();
